remove AbstractService.java, Move ClientMessageContext to Services package
