+++ server/src/main/java/org/schlunzis/kurtama/server/auth/AuthenticationService.java
@@ -12,9 +12,9 @@ import org.schlunzis.kurtama.common.messages.authentication.register.RegisterFai
 import org.schlunzis.kurtama.common.messages.authentication.register.RegisterRequest;
 import org.schlunzis.kurtama.common.messages.authentication.register.RegisterSuccessfulResponse;
 import org.schlunzis.kurtama.server.lobby.LobbyStore;
-import org.schlunzis.kurtama.server.net.ClientMessageContext;
 import org.schlunzis.kurtama.server.net.ISession;
 import org.schlunzis.kurtama.server.net.ServerMessageWrapper;
+import org.schlunzis.kurtama.server.service.ClientMessageContext;
 import org.schlunzis.kurtama.server.user.DBUser;
 import org.schlunzis.kurtama.server.user.IUserStore;
 import org.schlunzis.kurtama.server.user.ServerUser;
